// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/reviewStatus.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type ReviewStatus = 0 | 1 | 2;

export const ReviewStatus = {
    NUMBER_0: 0 as ReviewStatus,
    NUMBER_1: 1 as ReviewStatus,
    NUMBER_2: 2 as ReviewStatus
};